+++ front/src/components/RecipeList/RecipeCard/RecipeCard.module.scss
@@ -1,12 +1,21 @@
-.card {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  max-width: 632px;
-  height: 288px;
+.recipe {
+  
   border-radius: 16px;
   box-shadow: 3px 4px 24px 0px rgba(51, 45, 40, 0.06);
+
+  @media only screen and (min-width: 768px) {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    height: 288px;
+  }
+
+  @media only screen and (max-width: 767px) {
+    display: flex;
+    flex-direction: column;
+  }
 }
-.cardImage {
+
+.img {
   width: 100%;
   height: 100%;
   object-fit: cover;
@@ -15,8 +24,16 @@
     transition: opacity 0.4s linear, transform 0.4s linear, filter 0.3s linear;
   }
 }
-.cardName {
-  height: 165px;
+
+.description {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  padding: 20px 20px 20px 24px;
+}
+
+.title {
+  
   margin: 0 0 10px 0;
   padding: 0;
   font-family: 'Biolilbert', sans-serif;
@@ -27,20 +44,13 @@
   color: #232323;
   overflow: hidden;
   text-overflow: ellipsis;
+
+  @media only screen and (min-width: 768px) {
+    height: 165px;
+  }
 }
-.cardDescription {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  padding: 20px 20px 20px 24px;
-}
-.cardDescription__item {
-  display: flex;
-  margin-top: 12px;
-}
-.cardDescription__text {
-  margin: 0;
-  padding: 0;
+
+.complexity, .cookingTime {
   font-family: 'Mulish';
   font-size: 18px;
   font-style: normal;
@@ -48,7 +58,16 @@
   line-height: 22px;
   color: #909090;
 }
-.complexityImage {
+
+.complexity {
+  display: flex;
+  align-items: center;
+  margin-top: 0;
+  margin-bottom: 14px;
+}
+
+.star {
+  display: inline-block;
   width: 24px;
   height: 24px;
   margin-left: 8px;
@@ -58,12 +77,12 @@
   background-size: contain;
   background-repeat: no-repeat;
 }
-.timerImage {
-  width: 24px;
-  height: 24px;
-  margin-right: 4px;
+
+.cookingTime {
   background-image: url('../../../images/timer.svg');
-  background-position: 50%;
-  background-size: contain;
+  background-position: left center;
   background-repeat: no-repeat;
+  padding-left: 32px;
+  line-height: 24px;
+  margin: 0;
 }
