+++ front/src/components/RecipeList/RecipeList.module.scss
@@ -1,0 +1,27 @@
+.title {
+  margin-top: 0;
+  padding: 0;
+  font-family: 'Biolilbert', sans-serif;
+  font-size: 36px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 40px;
+  color: #232323;
+}
+
+.recipeList {
+  margin-top: 41px;
+  margin-bottom: 60px;
+
+  @media only screen and (min-width: 1231px) {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 35px;    
+  }
+
+  @media only screen and (max-width: 1230px) {
+    display: flex;
+    flex-direction: column;
+    gap: 35px;    
+  }
+}
